Feature: Future plan for the sidep roject and fix the foldering structure
