+++ stories/About.stories.mdx
@@ -2,7 +2,7 @@ import { Meta } from '@storybook/addon-docs';
 import { Dropdown } from '../src';
 import Header from './assets/ds-header.png';
 
-<Meta title="MDX/A. Introduction" component={Dropdown} />
+<Meta title="About" component={Dropdown} />
 
 # Welcome to MoonbladeUI
