+++ stories/Atomos/Atomos.stories.mdx
@@ -1,0 +1,43 @@
+import { Meta } from '@storybook/addon-docs';
+import AtomImg from '../assets/components/atom.png';
+
+<Meta title="Components/Atomos/Overview" />
+
+# Atomos
+
+<p>
+  <strong>Atomos</strong> is word derived from Greek work that's mean <strong>
+    Atom
+  </strong> . 
+
+  If atoms are the basic building blocks of matter, then the atoms of our interfaces serve as the foundational building blocks that comprise all our user interfaces. These atoms include basic HTML elements like form labels, inputs, buttons, and others that can’t be broken down any further without ceasing to be functional.
+</p>
+
+<br />
+<br />
+
+<div style={{ display: 'flex', flexDirection: 'column', justifyContent: 'center', alignItems: 'center' }}>
+  <img src={AtomImg} alt="Atom Image" />
+
+  <p>
+    Refence: <a href="https://atomicdesign.bradfrost.com/chapter-2/">Atomic Design</a>
+  </p>
+</div>
+<br />
+<br />
+
+## Upcoming Components and Enhancement
+
+<ul>
+  <li>Announcement Bar</li>
+  <li>Button & Links</li>
+  <li>Dots Indicator</li>
+  <li>Dropdown Menu</li>
+  <li>List</li>
+  <li>Loaders</li>
+  <li>Notice</li>
+  <li>Selection</li>
+  <li>Separators</li>
+  <li>States</li>
+  <li>Tooltips</li>
+</ul>
