+++ stories/ExistingToken.stories.mdx
@@ -2,7 +2,7 @@ import { Meta } from '@storybook/addon-docs';
 import { Dropdown } from '../src';
 import { Token } from "../src/Token";
 
-<Meta title="MDX/D. Existing Token" component={Dropdown} />
+<Meta title="Existing Token" component={Dropdown} />
 
 ## Custom Token For MoonbladeUI
