+++ stories/Installation.stories.mdx
@@ -4,7 +4,7 @@ import Sidebar from './assets/sidebar.png';
 import DropdownImg from './assets/dropdown.png';
 import SnippetCode from './assets/snippet-code.png';
 
-<Meta title="MDX/B. Installation" component={Dropdown} />
+<Meta title="Installation" component={Dropdown} />
 
 # Installation
