+++ stories/Mório/Mório.stories.mdx
@@ -1,0 +1,38 @@
+import { Meta } from '@storybook/addon-docs';
+import MoleculeImg from '../assets/components/molecule.png';
+
+<Meta title="Components/Mório/Overview" />
+
+# Mório
+
+<p> <strong>Mório</strong> is word derived from Greek work that's mean <strong>Molecules</strong> .
+
+In interfaces, <strong>molecules</strong> are relatively simple groups of UI elements functioning together as a unit. For example, a form label, search input, and button can join together to create a search form molecule.
+
+</p>
+<br/><br/>
+
+<div style={{ display: 'flex', flexDirection: 'column', justifyContent: 'center', alignItems: 'center' }}>
+  <img src={MoleculeImg} alt="Atom Image" />
+  
+  <p>
+    Refence: <a href="https://atomicdesign.bradfrost.com/chapter-2/">Atomic Design</a>
+  </p>
+</div>
+
+<br />
+<br />
+
+## Upcoming Components and Enhancement
+
+<ul>
+  <li>Social Sharing Bar</li>
+  <li>Section Header</li>
+  <li>Seach Bar</li>
+  <li>Pagination Bar</li>
+  <li>Filter</li>
+  <li>Cards</li>
+  <li>Breadcrumb</li>
+  <li>Alert</li>
+  <li>Accordion</li>
+</ul>
