+++ stories/Organismós/Organismós.stories.mdx
@@ -1,0 +1,24 @@
+import { Meta } from '@storybook/addon-docs';
+import OrganismImg from '../assets/components/organism.png';
+
+<Meta title="Components/Organismós/Overview" />
+
+# Organismós
+
+<p>
+  <strong>Organismós</strong> is word derived from Greek work that's mean <strong>Organism. </strong>
+
+  The <strong> Organismós are relatively complex UI components composed of groups of molecules and/or atoms and/or other organisms.</strong> These organisms form distinct sections of an interface.
+</p>
+<br />
+<br />
+
+<div style={{ display: 'flex', flexDirection: 'column', justifyContent: 'center', alignItems: 'center' }}>
+  <img src={OrganismImg} alt="Atom Image" />
+  <p>
+    Refence: <a href="https://atomicdesign.bradfrost.com/chapter-2/">Atomic Design</a>
+  </p>
+</div>
+
+<br />
+<br />
