+++ stories/Token.stories.mdx
@@ -2,7 +2,7 @@ import { Meta } from '@storybook/addon-docs';
 import { Dropdown } from '../src';
 import Header from './assets/ds-header.png';
 
-<Meta title="MDX/C. What is Token?" component={Dropdown} />
+<Meta title="What is Token?" component={Dropdown} />
 
 # Design Token
